% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{methods}
\name{associatedShortestPaths,Graph-method}
\alias{associatedShortestPaths,Graph-method}
\title{function that uses the object Graph and calculs distances for every pair of
gene - metabolite from data.}
\usage{
\S4method{associatedShortestPaths}{Graph}(object, data)
}
\arguments{
\item{object}{Graph, data(mg, m, genes, metabolites)}
}
\description{
Note that for 1 gene is attached 1 enzyme which is an edge in our graph.
In order to calcul a distance from vertice to vertice this function is
called twice in function 'getFinalDFSHortestDistance'.
}
\details{
First call is to calcul all the distance for the one vertice of the gene
to the all metabolites, the second call is to calcul the distance
from the other vertice to all metabolites and then
choose the smallest distance between the 2.

The selection of the first vertice and the second vertice is done in a other
function 'getIdGeneInGraph'

param data where
 mg = id (from igraph in object Graph) of 1 metabolite related to the gene
 m = id (igraph in object Graph)
 genes = hsa:... id from KEGG
 metabolites = C.... id from KEGG
}
\examples{
associatedShortestPaths(Graph, data)
calcul al distances
choosing smallest distance between the two metabolites of gene and
all metabolites
}
\keyword{kegg}

