% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotFunction.R
\name{barplot_adjustMaximalDistance}
\alias{barplot_adjustMaximalDistance}
\title{function to modify the frequency data.frame by adding more distance that are
then the one already defined}
\usage{
barplot_adjustMaximalDistance(maximumDistance, frequenciesDF,
  maxDistFrequenciesDF)
}
\description{
function to modify the frequency data.frame by adding more distance that are
then the one already defined
}

