% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{createGraphFromPathway}
\alias{createGraphFromPathway}
\title{create df for vertices
create df edges
create graphEl objects
create igraph with graphEl object elements
create Graph object}
\usage{
createGraphFromPathway(pathwayId)
}
\description{
create df for vertices
create df edges
create graphEl objects
create igraph with graphEl object elements
create Graph object
}

