% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{methods}
\name{createIGraph,GraphElements-method}
\alias{createIGraph,GraphElements-method}
\title{Creation of object igraph with data from XML file of pathway
of interest}
\usage{
\S4method{createIGraph}{GraphElements}(object)
}
\description{
Creation of object igraph with data from XML file of pathway
of interest
}

