% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{exportDFtoTxt}
\alias{exportDFtoTxt}
\title{Export results data of a data frame into a text files, for which each lines
separates the data from every data frame colum by tabulation. This text file
can be easily opened with excel.}
\usage{
exportDFtoTxt(export_data, exportNameFile)
}
\arguments{
\item{a}{dataframe.}
}
\description{
Export results data of a data frame into a text files, for which each lines
separates the data from every data frame colum by tabulation. This text file
can be easily opened with excel.
}
\examples{
exportDFtoTxt(dataFrame, exportFileName)
}
\keyword{data,}
\keyword{dataframe.}
\keyword{export,}
\keyword{txt,}
