% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeListFunctions.R
\name{finalReactionEdgeDF}
\alias{finalReactionEdgeDF}
\title{Since it is not possible to have the related gene when parsing the
reaction nodes in the KGML. I had to parse de reaction nodes and the
gene entry nodes separatly and then combinde the 2 with the information
needed.}
\usage{
finalReactionEdgeDF(pathwayId)
}
\arguments{
\item{id}{of the explored pathway}
}
\description{
the results gives a data.frame(substrateId, productId, subtrateName,
productName, reactionId, reactionName, ko(gene))
}
\examples{
finalReactionEdgeDF(pathwayId)
}
\keyword{data.frame}
\keyword{reaction}

