% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getCompoundNodeKgmlId}
\alias{getCompoundNodeKgmlId}
\title{A igraph function}
\usage{
getCompoundNodeKgmlId(g, compoundKeggId, nodeDF)
}
\arguments{
\item{graphe, }{kegg id of compound of interest, dataframe of compounds
(nodes)}
}
\description{
function returning the kgmlId of the compound of interest
}
\examples{
getCompoundNodeKgmlId(g, compoundKeggId, nodeDF)
}
\keyword{graph,}
\keyword{kgmlId}
\keyword{node,}

