% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundListFunctions.R
\name{getDataFrameOfDuplicateCompounds}
\alias{getDataFrameOfDuplicateCompounds}
\title{This fucntion creates a data frame which contains the duplicated compounds
in the initial graph. We need this because some reaction has a subtrate or
a product that has a metabolite that we find more than one in the graph but
we don't know which node we should use (reaction type ortholog). We prefer
not to use those reactions
#### ortholog}
\usage{
getDataFrameOfDuplicateCompounds(compoundDataFrame)
}
\arguments{
\item{compounds}{data frame}
}
\description{
This fucntion creates a data frame which contains the duplicated compounds
in the initial graph. We need this because some reaction has a subtrate or
a product that has a metabolite that we find more than one in the graph but
we don't know which node we should use (reaction type ortholog). We prefer
not to use those reactions
#### ortholog
}
\examples{
getDataFrameOfDuplicateCompounds(compoundDataFrame)
}
\keyword{compounds}

