% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAll}
\alias{getDistanceAll}
\title{Function calculating shortest distance between every gene in a
gene-metabolite pairs and all metabolite.}
\usage{
getDistanceAll(pathwayId, data, metabolite)
}
\arguments{
\item{pathwayId}{KEGG Id of selected map}

\item{data}{is a dataFrame with 2 columns. Where each line
reprensents an associations with the first column as
gene KEGG Ids and the sencond column as metabolite
KEGG Ids.}

\item{metabolite}{is a dataframe of 1 column with the KEGG Ids of
all measured metabolites.}
}
\description{
Function calculating shortest distance between every gene in a
gene-metabolite pairs of your data parameter and every metabolite
(in a pair or not) on a graph model of KEGG map selected, where nodes are
metabolites and reactions are edges.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.

Output : dataframe with metabolite in columns and gene in rows with
         shortest distance values.
}
\examples{
getDistanceAll(metabolismOverviewMapKEGGId,shinAndAlDF,
          completeMetaboDF)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

