% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAllMaps.R
\name{getDistanceAssoAllMaps}
\alias{getDistanceAssoAllMaps}
\title{isGeneInMap, metaboliteCommonName, metaboliteKEGGId,
isMetaboliteInMap, distance}
\usage{
getDistanceAssoAllMaps(association, ordered = FALSE, commonNames = TRUE)
}
\arguments{
\item{association}{Dataframe with 2 columns, where each line reprensents an
association. First column are the genes and the sencond column are the
metabolites. Only use KEGG Ids.}

\item{ordered}{[option] ascendent ordering of distance}

\item{commonNames}{get KEGG's Common Names of the KEGG Id in the
results.}
}
\description{
isGeneInMap, metaboliteCommonName, metaboliteKEGGId,
isMetaboliteInMap, distance
}
\examples{
getDistanceAssoAllMaps(shin,ordered = T, commonNames = F)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

