% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{methods}
\name{getFinalAssoDfSd,Graph-method}
\alias{getFinalAssoDfSd,Graph-method}
\title{A igraph function}
\usage{
\S4method{getFinalAssoDfSd}{Graph}(object, data)
}
\arguments{
\item{graphe, }{kegg id of compound of interest, dataframe of compounds
(nodes)}
}
\description{
Calls getIdGeneInGraph() twice to get :
Modification of dataframe giving in entry by user, from
c("gene", "metabolite") to 2 dataFrame
where indexMetabolite = 1
d1 = c("metabolite1giIdIniGraph", "metabolitesIdIniGraph")
#' where indexMetabolite = 2
d2 = c("metabolite2giIdIniGraph", "metabolitesIdIniGraph")
}
\details{
This is because genes are edges and we calculate vertex to vertex
the distance between genes and metabolites.
From each gene we want to take the shortest distance between the 2 vertex
attached, so we calcul both and in an other function we take the smallest
and that will be the output.

This function keeps, between the 2 metabolites of each gene vertexes,
the distance that is the smallest between the 2 to the same metabolite.
}
\examples{
getCompoundNodeKgmlId(g, data, indexMetabolite)
}
\keyword{igraph,}
\keyword{kgmlId}
\keyword{node,}

