% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotFunction.R
\name{getFrequenceAssociationsDF}
\alias{getFrequenceAssociationsDF}
\title{creation of vector to fill bar colors automatically}
\usage{
getFrequenceAssociationsDF(frequenceDistDF, shortestsPathsDF, gene)
}
\description{
creation of vector to fill bar colors automatically
}

