% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphFunctions.R
\name{getHeadTailKgmlIdOfEdge}
\alias{getHeadTailKgmlIdOfEdge}
\title{A igraph function}
\usage{
getHeadTailKgmlIdOfEdge(g, hsaGene, reactionDF)
}
\arguments{
\item{graphe, }{gene of interest, dataframe of reaction (edges) in the graph}
}
\description{
function returning the nodes (sub, prod ) of the egdes representing the gene
 of interest
}
\examples{
getHeadTailKgmlIdOfEdge(g, hsa, reactionDF)
}
\keyword{edge}
\keyword{graph,}

