% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{methods}
\name{getIdMetabolitesInGraph,Graph-method}
\alias{getIdMetabolitesInGraph,Graph-method}
\title{A igraph function}
\usage{
\S4method{getIdMetabolitesInGraph}{Graph}(object, completeMetaboDF)
}
\arguments{
\item{Graph, }{data, indexMetabolites}
}
\description{
where Graph param is the Graph object
where data is (gene, metabolite)
where indice metabolites is only 1 or 2
}
\examples{
getIdGeneInGraph(g, data, indexMetabolite)
}
\keyword{igraph,}
\keyword{kgmlId}
\keyword{node,}

