% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{getListEdgeFromGeneKGML}
\alias{getListEdgeFromGeneKGML}
\title{A KEGG Function}
\usage{
getListEdgeFromGeneKGML(pathwayId)
}
\arguments{
\item{pathway}{id}
}
\description{
This function extract from KGML a list of reaction of entry-type "gene"
Returns a dataFrame object : id, entryId, reaction, ko
}
\examples{
getListEdgeFromGeneKGML("hsa01100")
}
\keyword{Gene,}
\keyword{Reaction}
\keyword{kegg,}

