% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{getListNodeFromKGML}
\alias{getListNodeFromKGML}
\title{This function extract from KGML a list of metabolites (id and name) that
makes the nodes of the graph ** some metablite can from more than 1 node,
which is why the id of the nodes is not the name of the metabolite.
data.frame (id, keggId)}
\usage{
getListNodeFromKGML(pathwayId)
}
\arguments{
\item{pathway}{id}
}
\description{
This function extract from KGML a list of metabolites (id and name) that
makes the nodes of the graph ** some metablite can from more than 1 node,
which is why the id of the nodes is not the name of the metabolite.
data.frame (id, keggId)
}
\examples{
getListNodeFromKGML("hsa01100")
}
\keyword{kegg}

