% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseXML.R
\name{getListReactionFromKGML}
\alias{getListReactionFromKGML}
\title{A KEGGREST function}
\usage{
getListReactionFromKGML(pathwayId)
}
\arguments{
\item{pathway}{id}
}
\description{
This function creates a data.frame containing the first part of our edge list
the "gene" reactions
data.frame(kgmlIdSubstratre, kgmlIdProduct, substrateName, productName,
kgmlIdReaction, reactionName, reactionType)
}
\examples{
getListReactionFromKGML("hsa01100")
}
\keyword{keggrest,}
\keyword{kgml}

