% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getPath}
\alias{getPath}
\title{get the shortest path between on specific association between one gene and
one metabolite}
\usage{
getPath(pathwayId, gene, metabolite, commonNames = F)
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{gene}{KEGG Id.}

\item{commonNames}{get KEGG's Common Names of the KEGG Id provided in the
path results.}

\item{Metabolite}{KEGG Id.}
}
\description{
get the shortest path between on specific association between one gene and
one metabolite
}
\examples{
getPath("hsa:01100","hsa:1373", "C00300", T)
}
\keyword{association,}
\keyword{path.}

