% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractingKGMLfromKEGG.R
\name{getPathwayKGML}
\alias{getPathwayKGML}
\title{A KEGG Function}
\usage{
getPathwayKGML(pathwayId)
}
\arguments{
\item{this}{function parses all data from kegg}
}
\description{
This function allows you to get a KGML format of a pathway and save
xml format in data directory
}
\examples{
getPathwayKGML("hsa01100")
}
\keyword{kegg}

