% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{class}
\name{Graph-class}
\alias{Graph-class}
\title{S4 classe Graph, contains all information important to create graphe argument
and to calcul distances for it}
\description{
S4 classe Graph, contains all information important to create graphe argument
and to calcul distances for it
}

