% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\docType{class}
\name{igraph-class}
\alias{igraph-class}
\title{object igraph instanciated as S4 classe to be able to use igraph objects in
other S4 classes (Graphe)}
\description{
object igraph instanciated as S4 classe to be able to use igraph objects in
other S4 classes (Graphe)
}

