% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{inputXLSXtoDF}
\alias{inputXLSXtoDF}
\title{Inporting data from an XSLX file, in the context of this package, one of the
data to import is the association file.}
\usage{
inputXLSXtoDF(association)
}
\arguments{
\item{path}{to XSLX file with 2 columns where each line reprensents an
association. First column are genes and the sencond column are
metabolites. Only use KEGG Ids.}
}
\description{
Inporting data from an XSLX file, in the context of this package, one of the
data to import is the association file.
}
\examples{
inputXLSXtoDF("file_path/fileName.xlsx")
}
\keyword{XSLX,}
\keyword{associations.}
\keyword{data}
\keyword{data,}
\keyword{frame,}
\keyword{input,}

