% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameFunctions.R
\name{mergeRowsWithSmallestValueByKEGGId}
\alias{mergeRowsWithSmallestValueByKEGGId}
\title{merge rows with smallest values in each col for repeated metabolites or genes
Genes and metabolites are likely to be repeated since we can find them in
multiple places in the graph reprensentation of a map}
\usage{
mergeRowsWithSmallestValueByKEGGId(df)
}
\description{
merge rows with smallest values in each col for repeated metabolites or genes
Genes and metabolites are likely to be repeated since we can find them in
multiple places in the graph reprensentation of a map
}

