% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{mergeVectorsLowerValues}
\alias{mergeVectorsLowerValues}
\title{function merge 2 numeric vectors and return a vector with the smalest
values for each position of the vectors}
\usage{
mergeVectorsLowerValues(A, B)
}
\arguments{
\item{2}{numeric vectors}
}
\description{
function merge 2 numeric vectors and return a vector with the smalest
values for each position of the vectors
}
\examples{
d1 <- c(1,2,1,2,3,4), d2 <- c(2,3,1,1,5,2)
 mergeVectorsLowerValues(d1,d2) returns -> 1,2,1,1,3,2
}
\keyword{kegg}

