% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameFunctions.R
\name{removeRowsDistanceAll}
\alias{removeRowsDistanceAll}
\title{Thsi function also sorts the data by metabolite or gene ids... we don't want
that}
\usage{
removeRowsDistanceAll(df)
}
\description{
Thsi function also sorts the data by metabolite or gene ids... we don't want
that
}

