% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeListFunctions.R
\name{unlistEdgeReactionCol}
\alias{unlistEdgeReactionCol}
\title{Function fractioning reaction list from every gene from KGML file}
\usage{
unlistEdgeReactionCol(edgeDataFrame)
}
\arguments{
\item{data}{frame des reactions extraites du KGML en listes}
}
\description{
exemple : id R1... R2... R3... ko
-> id R1 ko
-> id R2 ko
-> id R3 ko
}
\examples{
completeEdgeList(edgeDataFrame)
}
\keyword{kegg}

