% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotFunction.R
\name{distanceGeneToAllMetabolite}
\alias{distanceGeneToAllMetabolite}
\title{function that output a barplot graph related to one specific gene with all
the shortest distances from that gene to all metabolites}
\usage{
distanceGeneToAllMetabolite(pathwayId, associatedGeneMetaDF,
  completeMetaboliteDF, gene)
}
\arguments{
\item{KEGGPathwayId, }{data(gene, metabolites),allMeasuredMetaboliteesDF, gene}
}
\description{
for param data:
     gene = KEGGid of gene hsa:...
     metabolites : KEGGid of metabolites C....
for param pathwayId : KEGG id of pathways without ':' ex: hsa01100
for param gene : is a gene in data ex: hsa:8801
}
\examples{
barplotFunction(hsa01100, associatedGeneMetaDF,
 completeMetaboliteDF, gene)
get all shortest paths from data entry
adjust gene parameter
add metabolite row
get a subset of shortestsPathsDF contaning only geneOf interest, gene
and metaboltie column
get frequency of every value until the maxVal found + Inf val
set column names of frequenceDF
get associations values for all distance pair calculated
bind info into 1 DF
Association of the distance where there is an association
}
\keyword{Graph,}
\keyword{barplot,}
\keyword{shortestDistance}

