% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotFunction.R
\name{distributionGene}
\alias{distributionGene}
\title{Function calculating shortest distance between a selected gene of a
gene-metabolite pairs and every metabolites.}
\usage{
distributionGene(pathwayId, data, metabolite, gene)
}
\arguments{
\item{pathwayId}{KEGG Id of selected map}

\item{data}{is a dataFrame with 2 columns. Where each line
reprensents an associations with the first column as
gene KEGG Ids and the sencond column as metabolite
KEGG Ids.}

\item{metabolite}{is a dataframe of 1 column with the KEGG Ids of
all measured metabolites.}

\item{gene}{is the KEGG Id of the selected gene}
}
\description{
Function calculating shortest distance between a selected gene of a
gene-metabolite pairs and every metabolites (in a pair or not)
on a graph model of KEGG map selected, where nodes are metabolites and
reactions are edges.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.

Output: barplot showing the distribution of the calculated distance, where
red bars represent a distance with an associated metabolite and grey bars
distance with no associated metabolites.
}
\examples{
distributionGene(metabolismOverviewMapKEGGId,
shinAndAlDF, completeMetaboDF, "hsa:1373")
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

