% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotFunction.R
\name{distributionGene}
\alias{distributionGene}
\title{Distance distribution plots for single gene}
\usage{
distributionGene(pathwayId, association, metabolite, gene)
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{association}{Dataframe with 2 columns, where each line reprensents an
associations. First column are the genes and the sencond column as the
metabolites. Only use KEGG Ids.}

\item{metabolite}{Dataframe of 1 column, representing all the measured
metabolites. Only use KEGG Ids.}

\item{gene.}{Gene selected Only use KEGG.}
}
\description{
Function ploting the distribution of distances between a gene and all
measured metabolite, highlithing the distance of its associated metabolites.
}
\details{
The plot is depicted as frequency bars, which represent
the number of metabolites at a given distance for the selected gene.
Frequency bars are shown in grey for metabolites that are not associated with
the selected gene and in red if there is at least one metabolite associated
with this gene.

If a gene or a metabolite is present on multiple edges or nodes, then the
shortest distance is selected.
}
\examples{
distributionGene(metabolismOverviewMapKEGGId,
shinAndAlDF, completeMetaboDF, "hsa:1373")
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

