% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAll}
\alias{getDistanceAll}
\title{Function calculating shortest distance between every gene and every
metabolite.}
\usage{
getDistanceAll(pathwayId, gene, metabolite)
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{gene}{Dataframe of 1 column, representing all genes reported. Only use
KEGG Ids.}

\item{metabolite}{Dataframe of 1 column, representing all the measured
metabolites. Only use KEGG Ids.}
}
\description{
Function calculating shortest distance between every gene and every
metabolites on a graph model of KEGG map selected, where nodes are
metabolites and reactions (genes) are edges.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.

Output : dataframe with metabolite in columns and gene in rows with
         shortest distance values.
}
\examples{
getDistanceAll(metabolismOverviewMapKEGGId,completeGeneDF,
          completeMetaboDF)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}
