% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAll}
\alias{getDistanceAll}
\title{Fonction that calculates every shortest distances between each gene and
all metabolites in KEGG pathway of choice}
\usage{
getDistanceAll(pathwayId, associatedGeneMetaDF, completeMetaboliteDF)
}
\arguments{
\item{data(gene, }{metabolites )}
}
\description{
The igrpah created to simulate the KEGG pathways as metabolties as nodes
and genes (related gene enzymes and reaction) as egdes.
}
\details{
The shortest distance is taking from calculation from both vertices related
to a gene to the metabolites of interest.

for param data:
     gene = KEGGid of gene hsa:...
     metabolites : KEGGid of metabolites C....
for param pathwayId : KEGG id of pathways without ':' ex: hsa01100
}
\examples{
getDistanceAll(pathwayId, data)
}
\keyword{KEGG}

