% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAsso}
\alias{getDistanceAsso}
\title{Function calculating shortest distance between each gene-metabolite pairs.}
\usage{
getDistanceAsso(pathwayId, data, ordered = FALSE)
}
\arguments{
\item{pathwayId}{KEGG Id of selected map}

\item{data}{is a dataFrame with 2 columns. Where each line
reprensents an associations with the first column as
gene KEGG Ids and the sencond column as metabolite
KEGG Ids.}

\item{ordered}{[option] ascendent ordering of distance}
}
\description{
Function calculating shortest distance between each gene-metabolite pairs of
your data parameter on a graph model of KEGG map selected, where nodes are
metabolites and reactions are edges.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.

Output : dataframe with the following columns : geneCommonName, geneKEGGId,
isGeneInMap, metaboliteCommonName, metaboliteKEGGId,
isMetaboliteInMap, distance
}
\examples{
getDistanceAsso(metabolismOverviewMapKEGGId,shinAndAlDF)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

