% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAsso}
\alias{getDistanceAsso}
\title{Function calculating shortest distance between each gene-metabolite pairs.}
\usage{
getDistanceAsso(pathwayId, association, ordered = FALSE, commonNames = TRUE)
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{association}{Dataframe with 2 columns, where each line reprensents an
association. First column are the genes and the sencond column are the
metabolites. Only use KEGG Ids.}

\item{ordered}{[option] ascendent ordering of distance}

\item{commonNames}{get KEGG's Common Names of the KEGG Id in the
results.}
}
\description{
Function calculating shortest distance between each gene-metabolite
associations on your selected KEGG pathway.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then the
shortest distance is selected.
Output : dataframe with the following columns : geneCommonName, geneKEGGId,
isGeneInMap, metaboliteCommonName, metaboliteKEGGId,
isMetaboliteInMap, distance
}
\examples{
getDistanceAsso("hsa01100",shinAndAlDF)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{shortestDistance,}

