% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{getDistanceAsso}
\alias{getDistanceAsso}
\title{Fonction that calculates distance between each gene-metabolite pairs.}
\usage{
getDistanceAsso(pathwayId, data, ordered = FALSE, output = c("xslx",
  "data.frame"))
}
\arguments{
\item{data(gene, }{metabolites )}
}
\description{
The igrpah created to simulate the KEGG pathways as metabolties as nodes
and genes (related gene enzymes and reaction) as egdes.
}
\details{
The shortest distance is taking from calculation from both vertices related
to a gene to the metabolites of interest.

for param data:
     gene = KEGGid of gene hsa:...
     metabolites : KEGGid of metabolites C....
for param pathwayId : KEGG id of pathways without ':' ex: hsa01100
}
\examples{
getDistanceAsso(pathwayId, data, ouput)
}
\keyword{KEGG}

