% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFunction.R
\name{heatmapAsso}
\alias{heatmapAsso}
\title{Function that output a heatmap to visualize distance calculated between every
gene-metabolite associations in input.}
\usage{
heatmapAsso(pathwayId, data)
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{association}{Dataframe with 2 columns, where each line reprensents an
associations. First column are the genes and the sencond column as the
metabolites. Only use KEGG Ids.}
}
\description{
Function calculting shortest distance between every genes and metabolites in
a gene-metabolite pairs of your data parameter on a graph model of KEGG map
selected, where nodes are metabolites and reactions are edges.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.

Output : Heatmap graphic where black countoured distances represent
          an association.
}
\examples{
heatmapAsso(metabolismOverviewMapKEGGId, shinAndAlDF)
}
\keyword{KEGG}
\keyword{graph,}
\keyword{heatmap,}
\keyword{shortestDistance,}

