% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationFunction.R
\name{permutationTest}
\alias{permutationTest}
\title{Statistical permutation test to asses the relevance of associated
gene-metabolite pairs vs. randomly selected pairs}
\usage{
permutationTest(pathwayId, data, gene, metabolite, permutation,
  output = c("medians", "pvalue", "histogram"))
}
\arguments{
\item{pathwayId}{KEGG Id of selected map}

\item{data}{is a dataframe with 2 columns. Where each line
reprensents an associations with the first column as
gene KEGG Ids and the sencond column as metabolite
KEGG Ids.}

\item{gene}{is a dataframe of 1 column with the KEGG Ids of
all measured genes.}

\item{metabolite}{is a dataframe of 1 column with the KEGG Ids of
all measured metabolites.}

\item{permutation}{is the number desired permutations}

\item{output}{medians (list of median of every permutation), pvalue of the
permutation test, histogram (a histrogram representing the
distribution of every permutation median)}
}
\description{
Permutation test to evauluate if gene-metabolite associations pairs of
your data parameter are significantly closer than randomly selected
gene-metabolite pairs by calculating shortest distance between assocatied
gene-metabolite pairs vs. randomly selected pairs.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then
shortest distance are calculated for every combinaison possible and the
shortest distance is selected.
}
\examples{
permutationFunction(metabolismOverviewMapKEGGId, shinAndAlDF,
            completeGeneDF,completeMetaboDF, 1000, "histogram")
}
\keyword{graph,KEGG}
\keyword{median,}
\keyword{permutation,}
\keyword{shrotestDistance,}
\keyword{statistic}
\keyword{test,}

