% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationFunction.R
\name{permutationTest}
\alias{permutationTest}
\title{Statistical permutation test to asses the relevance of associated
gene-metabolite pairs vs. randomly selected pairs}
\usage{
permutationTest(pathwayId, association, gene, metabolite, permutation,
  output = c("medians", "pvalue", "histogram"))
}
\arguments{
\item{pathwayId}{KEGG Id of selected pathway.}

\item{association}{Dataframe with 2 columns, where each line reprensents an
association. First column are genes and the sencond column are
metabolites. Only use KEGG Ids.}

\item{gene}{Dataframe of 1 column, representing all genes reported. Only use
KEGG Ids.}

\item{metabolite}{Dataframe of 1 column, representing all the measured
metabolites. Only use KEGG Ids.}

\item{permutation}{Number desired permutations}

\item{output}{'medians' of all permutations, 'pvalue' of the
permutation test, 'histogram' representing the
distribution of all permutations' median}
}
\description{
Permutation test to evauluate if gene-metabolite associations are significantly closer than randomly selected
gene-metabolite pairs.
}
\details{
If a gene or a metabolite is present on multiple edges or nodes, then the shortest distance is selected.
}
\examples{
permutationFunction("hsa01100", shinAndAlDF,
            completeGeneDF,completeMetaboDF, 1000, "histogram")
}
\keyword{KEGG}
\keyword{graph,}
\keyword{median,}
\keyword{permutation,}
\keyword{shrotestDistance,}
\keyword{statistic}
\keyword{test,}
